feat: add more error telemetry dimensions (#9598)
* feat: add more error telemetry dimensions

* feat: add inner error code

* feat: add inner error code

* fix: cycle import

* fix: error

* fix: error

* fix: error

* test: ut

* test: ut
