+++ packages/fx-core/src/common/m365/packageService.ts
@@ -11,6 +11,8 @@ import { waitSeconds } from "../tools";
 import { CoreSource } from "../../core/error";
 import { NotExtendedToM365Error } from "./errors";
 import { assembleError } from "../../error/common";
+import { ErrorContextMW } from "../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 // Call m365 service for package CRUD
 export class PackageService {
@@ -25,7 +27,7 @@ export class PackageService {
     this.initEndpoint = endpoint;
     this.logger = logger;
   }
-
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   private async getTitleServiceUrl(token: string): Promise<string> {
     try {
       const envInfo = await this.axiosInstance.get("/config/v1/environment", {
@@ -42,6 +44,7 @@ export class PackageService {
     }
   }
 
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   public async sideLoading(token: string, manifestPath: string): Promise<[string, string]> {
     try {
       const data = await fs.readFile(manifestPath);
@@ -111,7 +114,7 @@ export class PackageService {
       throw assembleError(error, CoreSource);
     }
   }
-
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   public async getLaunchInfoByManifestId(token: string, manifestId: string): Promise<any> {
     try {
       const serviceUrl = await this.getTitleServiceUrl(token);
@@ -164,7 +167,7 @@ export class PackageService {
       throw assembleError(error, CoreSource);
     }
   }
-
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   public async retrieveTitleId(token: string, manifestId: string): Promise<string> {
     const launchInfo = await this.getLaunchInfoByManifestId(token, manifestId);
     const titleId =
@@ -180,7 +183,7 @@ export class PackageService {
     this.logger?.debug(`AppId: ${appId}`);
     return appId;
   }
-
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   public async unacquire(token: string, titleId: string): Promise<void> {
     try {
       const serviceUrl = await this.getTitleServiceUrl(token);
@@ -204,7 +207,7 @@ export class PackageService {
       throw assembleError(error, CoreSource);
     }
   }
-
+  @hooks([ErrorContextMW({ source: "M365", component: "PackageService" })])
   public async getLaunchInfoByTitleId(token: string, titleId: string): Promise<unknown> {
     try {
       const serviceUrl = await this.getTitleServiceUrl(token);
