+++ packages/fx-core/src/component/driver/arm/deployImpl.ts
@@ -25,6 +25,8 @@ import { ensureBicepForDriver } from "./util/bicepChecker";
 import { ArmErrorHandle, DeployContext } from "./util/handleError";
 import { convertOutputs, getFileExtension, hasBicepTemplate } from "./util/util";
 import { validateArgs } from "./validator";
+import { ErrorContextMW } from "../../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 const helpLink = "https://aka.ms/teamsfx-actions/arm-deploy";
 
@@ -141,7 +143,7 @@ export class ArmDeployImpl {
       return errRes;
     }
   }
-
+  @hooks([ErrorContextMW({ source: "Azure", component: "ArmDeployImpl" })])
   async innerExecuteDeployment(
     templateArg: templateArgs,
     deploymentParameters: Deployment
