+++ packages/fx-core/src/component/driver/deploy/azure/impl/AzureZipDeployImpl.ts
@@ -23,6 +23,8 @@ import { getLocalizedString } from "../../../../../common/localizeUtils";
 import path from "path";
 import { zipFolderAsync } from "../../../../utils/fileOperation";
 import { DeployZipPackageError } from "../../../../../error/deploy";
+import { ErrorContextMW } from "../../../../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 export class AzureZipDeployImpl extends AzureDeployImpl {
   pattern =
@@ -152,6 +154,7 @@ export class AzureZipDeployImpl extends AzureDeployImpl {
    * @param logger log provider
    * @protected
    */
+  @hooks([ErrorContextMW({ source: "Azure", component: "AzureZipDeployImpl" })])
   async zipDeployPackage(
     zipDeployEndpoint: string,
     zipBuffer: Buffer,
