+++ packages/fx-core/src/component/resource/botService/appStudio/appStudioClient.ts
@@ -25,6 +25,8 @@ import {
 import { Messages } from "../messages";
 import { RetryHandler } from "../retryHandler";
 import { CommonStrings, ConfigNames } from "../strings";
+import { ErrorContextMW } from "../../../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 function handleBotFrameworkError(e: any, apiName: string): void | undefined {
   if (e.response?.status === HttpStatusCode.NOTFOUND) {
@@ -72,7 +74,7 @@ export class AppStudioClient {
   public static setRegion(region: string) {
     AppStudioClient.baseUrl = region;
   }
-
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async getBotRegistration(
     token: string,
     botId: string
@@ -96,7 +98,7 @@ export class AppStudioClient {
       handleBotFrameworkError(e, APP_STUDIO_API_NAMES.GET_BOT);
     }
   }
-
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async createBotRegistration(
     token: string,
     registration: IBotRegistration,
@@ -149,7 +151,7 @@ export class AppStudioClient {
 
     return;
   }
-
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async updateBotRegistration(
     token: string,
     botReg: IBotRegistration
