+++ packages/fx-core/src/core/globalVars.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { HookContext, Middleware, NextFunction } from "@feathersjs/hooks";
 import { Tools } from "@microsoft/teamsfx-api";
 
 export let TOOLS: Tools;
@@ -19,5 +20,57 @@ class GlobalVars {
   trackingId?: string;
   ymlFilePath?: string;
   envFilePath?: string;
+
+  //properties for error telemetry
+  stage = "";
+  source: ExternalSource = "";
+  component = "";
+  method = "";
 }
 export const globalVars = new GlobalVars();
+
+export interface ErrorContextOption {
+  component?: string;
+  source?: ExternalSource;
+  stage?: string;
+  method?: string;
+  reset?: boolean;
+}
+
+export function ErrorContextMW(option: ErrorContextOption): Middleware {
+  return async (ctx: HookContext, next: NextFunction) => {
+    option.method = ctx.method;
+    setErrorContext(option);
+    await next();
+    if (option.reset) {
+      resetErrorContext();
+    }
+  };
+}
+
+export function resetErrorContext(): void {
+  globalVars.stage = "";
+  globalVars.source = "";
+  globalVars.component = "";
+  globalVars.method = "";
+}
+
+export function setErrorContext(option: ErrorContextOption): void {
+  if (option.reset) {
+    resetErrorContext();
+  }
+  globalVars.component = option.component || globalVars.component;
+  globalVars.source = option.source || globalVars.source;
+  globalVars.stage = option.stage || globalVars.stage;
+  globalVars.method = option.method || globalVars.method;
+}
+
+export type ExternalSource =
+  | "Graph"
+  | "Azure"
+  | "Teams"
+  | "BotFx"
+  | "SPFx"
+  | "DevTools"
+  | "M365"
+  | "";
